feat: ensure ChangeProfilePicture calls DeleteFile when file does not exists and SaveUserPictureRepo throws

+++ src/data/use-cases/change-profile-picture.ts
@@ -23,7 +23,10 @@ export const setupChangeProfilePicture: Setup = (fileStorage, crypto, userProfil
     try {
       await userProfileRepo.savePicture(userProfile);
     } catch {
-      await fileStorage.delete({ key });
+      if (file) {
+        await fileStorage.delete({ key });
+      }
+      throw new Error();
     }
     return userProfile;
   };
